src/telegram-bot/requirements.txt
python-telegram-bot==21.11.1
confluent-kafka==2.8.0
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
structlog==25.1.0
pytz==2025.2
